Fix ESP32 limit JSON format types

+++ firmware/components/web_server/web_server_adapter_json.c
@@ -79,12 +79,15 @@ app_error_code_t web_adapter_build_limits_json(char *output, size_t output_size)
     }
     const int length =
         snprintf(output, output_size,
-                 "{\"ok\":true,\"data\":{\"macroNameBytes\":%u,\"macroSourceBytes\":%u,"
-                 "\"compiledActions\":%u,\"delayMs\":%u,\"durationMs\":%u,"
-                 "\"macrosPerPackage\":%u,\"packages\":%u,\"importBytes\":%u}}",
-                 APP_MACRO_NAME_MAX_BYTES, APP_MACRO_SOURCE_MAX_BYTES, APP_COMPILED_ACTION_MAX,
-                 APP_DELAY_MAX_MS, APP_ESTIMATED_DURATION_MAX_MS, APP_MACROS_PER_SET_MAX,
-                 APP_MACRO_SETS_MAX, APP_IMPORT_PACKAGE_MAX_BYTES);
+                 "{\"ok\":true,\"data\":{\"macroNameBytes\":%lu,\"macroSourceBytes\":%lu,"
+                 "\"compiledActions\":%lu,\"delayMs\":%lu,\"durationMs\":%lu,"
+                 "\"macrosPerPackage\":%lu,\"packages\":%lu,\"importBytes\":%lu}}",
+                 (unsigned long)APP_MACRO_NAME_MAX_BYTES,
+                 (unsigned long)APP_MACRO_SOURCE_MAX_BYTES,
+                 (unsigned long)APP_COMPILED_ACTION_MAX, (unsigned long)APP_DELAY_MAX_MS,
+                 (unsigned long)APP_ESTIMATED_DURATION_MAX_MS,
+                 (unsigned long)APP_MACROS_PER_SET_MAX, (unsigned long)APP_MACRO_SETS_MAX,
+                 (unsigned long)APP_IMPORT_PACKAGE_MAX_BYTES);
     if (length < 0 || (size_t)length >= output_size) {
         output[0] = '\0';
         return APP_ERROR_INTERNAL;
